# service_analytic/virtual-assistant-back/storage/api-docs/openapi.yaml
openapi: 3.0.2
info:
  title: API Виртуальный помощник
  description: 'API проекта Виртуальный помощник. Если конечная точка защищена, то может возвращаться ошибка с кодом 401'
  version: 2.0.0
servers:
  - url: 'https://test-vpcity.itech-test.ru/backend/api/v2'
    description: test server
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: ''
...
  schemas:
    Product:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        category:
          type: string
        position:
          type: number
        rating:
          type: number
        reviews:
          type: number
        characteristics:
          type: string
        updated_at:
          type: string
        price:
          type: number
        product_ozon_url:
          type: string
        is_advertised:
          type: boolean
        triggers:
          type: object
          properties:
            characteristics:
              type: integer
            position:
              type: boolean
            reviews:
              type: boolean
      required:
        - id
        - name
        - category
        - position
        - rating
        - reviews
        - characteristics
        - updated_at
        - price
        - product_ozon_url
        - triggers
paths:
...
  /products:
    get:
      tags:
        - OzProduct
      description: Получение таблицы товаров с сортировкой и пагинацией
      operationId: product-list
      security:
        - bearer: [ ]
      parameters:
        - schema:
            type: string
            default: application/json
          in: header
          name: content-type
          required: true
        - name: search
          in: query
          description: Поиск по строке (полям name, sku, id, external_id) всегда используется LIKE (кроме поля id)
          schema:
            type: string
            example: шнур
        - name: sortType
          in: query
          description: Тип сортировки desc или asc. Если параметр не указан, используется asc.
          schema:
            type: string
            example: desc
        - name: sortBy
          in: query
          description: Поле, по которому проходит сортировка
          schema:
            type: string
            example: price
        - name: page
          in: query
          description: Номер страницы. Если номер страницы больше максимального, устанавливается максимальное значение.
          schema:
            type: string
            example: 1
        - name: perPage
          in: query
          description: Количество товаров на странице. По умолчанию 100
          schema:
            type: string
            example: 100
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Product'
                      current_page:
                        type: integer
                      per_page:
                        type: integer
                      total:
                        type: integer
                      last_page:
                        type: integer
                      show_advert:
                        type: boolean
                      loaded_products:
                        type: object
                        properties:
                          code:
                            type: integer
                          status:
                            type: string
                      temp_products:
                        type: boolean
              examples:
                Example:
                  value:
                    data:
                      items:
                        - id: 2,
                          name: Marvy Uchida Набор линеров-ручек для черчения и рисования (темно-коричневый, сепия) 3 шт в упаковке, 0.03 мм,
                          category: Ручка капиллярная,
                          position: null
                          rating: 2,
                          reviews: 4,
                          characteristics: 9 из 37
                          update_at: 2021-03-19T12:14:05.000000Z
                          price: 383,
                          product_ozon_url: https://www.ozon.ru/context/detail/id/241793571
                          triggers:
                            characteristics: false
                            position: 2
                            reviews: false
                          is_advertised: 1
                        - id: 3,
                          name: Marvy Uchida Набор линеров-ручек для черчения и рисования (темно-коричневый, сепия) 5 шт в упаковке, 0.03 мм,
                          category: Ручка капиллярная,
                          position: null
                          rating: 2,
                          reviews: 4,
                          characteristics: 7 из 37
                          update_at: 2021-03-19T12:14:05.000000Z
                          price: 383,
                          product_ozon_url: https://www.ozon.ru/context/detail/id/241793571
                          triggers:
                            characteristics: false
                            position: -2
                            reviews: true
                          is_advertised: 0
                      current_page: 1,
                      per_page: 100,
                      total: 2,
                      last_page: 1
                      temp_products: true
                      loaded_products:
                        - sku: 248968180,
                          code: 200
                          status: Успешно добавлен
                        - sku: 248968181,
                          code: 3002
                          status: Товар был добавлен ранее
                        - sku: 248968181,
                          code: 3010
                          status: Товар добавлен в систему другим пользователем
                        - sku: 248968189,
                          code: 3000
                          status: Товар не продается
                        - sku: 248968183,
                          code: 400
                          status: Неверный Артикул
                        - sku: 24896818,
                          code: 404
                          status: Ошибка получения данных о товаре. Проверьте, есть ли в вашей учётной записи Озон данный товар
        '401':
          $ref: '#/components/responses/401'
...